REM-899 - Finish new sync implementation

+++ sync/src/main/kotlin/com/github/naz013/sync/usecase/DeleteSingleUseCase.kt
@@ -1,21 +1,24 @@
 package com.github.naz013.sync.usecase
 
-import com.github.naz013.cloudapi.CloudFileApi
-import com.github.naz013.domain.sync.SyncState
 import com.github.naz013.repository.RemoteFileMetadataRepository
+import com.github.naz013.sync.CloudApiProvider
 import com.github.naz013.sync.DataType
-import com.github.naz013.sync.SyncDataConverter
-import com.github.naz013.sync.local.DataTypeRepositoryCallerFactory
 
 internal class DeleteSingleUseCase(
-  private val cloudFileApi: CloudFileApi,
   private val remoteFileMetadataRepository: RemoteFileMetadataRepository,
-  private val getCloudFileNameUseCase: GetCloudFileNameUseCase
+  private val getCloudFileNameUseCase: GetCloudFileNameUseCase,
+  private val cloudApiProvider: CloudApiProvider,
 ) {
   suspend operator fun invoke(dataType: DataType, id: String) {
+    if (dataType == DataType.Settings) {
+      // Settings are not deleted from cloud.
+      return
+    }
     try {
       val fileName = getCloudFileNameUseCase(dataType, id)
-      cloudFileApi.deleteFile(fileName)
+      cloudApiProvider.getAllowedCloudApis().forEach { cloudFileApi ->
+        cloudFileApi.deleteFile(fileName)
+      }
       remoteFileMetadataRepository.deleteByLocalUuId(id)
     } catch (e: Exception) {
       throw e
